Upgrade episode page HTML.
